move TestMemRowNumber to row_number_test.go
